klasor yapisi duzenlendi ve java dosyasindan deilde propertiesden alinmasi saglandi
