+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+myapp.mail="your-stmp-gmail-address" 
+myapp.password="your-stmp-password"
